+++ main.c
@@ -6,118 +6,11 @@
 /*   By: takkatao <takkatao@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/02 16:26:14 by ahayashi          #+#    #+#             */
-/*   Updated: 2022/05/29 16:12:51 by takkatao         ###   ########.fr       */
+/*   Updated: 2022/05/31 14:42:39 by takkatao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
-
-/*
- * カメラから見る方向への、正規化されたrayを計算する
- * カメラの位置(lookfrom)は入力ファイルのview_pointから設定。
- * カメラの見る向きは入力ファイルのorientation_vecの向きに設定。
- */
-t_vec3	*calc_camera_ray(t_scene *scene)
-{
-	t_vec3	*lookat;
-	t_vec3	*lookfrom;
-	t_vec3	*camera_ray;
-	t_vec3	*camera_ray_norm;
-
-	lookfrom = scene->camera->point;
-	lookat = vec3_add(lookfrom, scene->camera->orientation_vec);
-	camera_ray = vec3_sub(lookat, lookfrom);
-	camera_ray_norm = vec3_normalize(camera_ray);
-	free(lookat);
-	free(camera_ray);
-	return (camera_ray_norm);
-}
-
-/*
- * vupを計算する。vupはカメラの上方向のベクトル。
- * カメラはz軸の正方向を上にしているとする。
- * カメラの視線ベクトルとvupが並行になってしまうと外積が0になってしまうため、少し角度をつける。
- */
-t_vec3	*calc_vup(t_vec3 *camera_ray)
-{
-	double	sum;
-
-	if (camera_ray->x == 0 && camera_ray->y == 0)
-	{
-		sum = sqrt(pow(0.1, 2) + 1);
-		return (vector3(0, 0.1 / sum, 1 / sum));
-	}
-	return (vector3(0, 0, 1));
-}
-
-t_vec3	*calc_u_base(t_vec3	*camera_ray)
-{
-	t_vec3	*u_base;
-	t_vec3	*vup;
-	t_vec3	*tmp1;
-
-	vup = calc_vup(camera_ray);
-	tmp1 = vec3_outer_product(vup, camera_ray);
-	u_base = vec3_normalize(tmp1);
-	free(vup);
-	free(tmp1);
-	return (u_base);
-}
-
-t_vec3	*calc_u(t_scene *scene, t_vec3	*camera_ray, double x)
-{
-	t_vec3	*u;
-	t_vec3	*u_base;
-	double	screen_width;
-
-	screen_width = tan(M_PI * scene->camera->fov / 180);
-	u_base = calc_u_base(camera_ray);
-	u = vec3_multiply(u_base, screen_width - 2 * screen_width * x / W);
-	free(u_base);
-	return (u);
-}
-
-t_vec3	*calc_v(t_scene *scene, t_vec3	*camera_ray, double y)
-{
-	t_vec3	*v;
-	t_vec3	*v_base;
-	t_vec3	*u_base;
-	t_vec3	*tmp1;
-	double	screen_width;
-
-	screen_width = tan(M_PI * scene->camera->fov / 180);
-	u_base = calc_u_base(camera_ray);
-	tmp1 = vec3_outer_product(u_base, camera_ray);
-	v_base = vec3_normalize(tmp1);
-	v = vec3_multiply(v_base, -1 * screen_width + 2 * screen_width * y / H);
-	free(tmp1);
-	free(u_base);
-	free(v_base);
-	return (v);
-}
-
-/*
- * スクリーン座標をワールド座標に変換。
- * minilibXのpixelを3次元上のスクリーンに変換。
- * カメラからスクリーンの距離は1で固定。
- * u,vはスクリーン上での基底ベクトル。基底ベクトルはvupと直交するように設定する。
- */
-t_vec3	*to_3d(t_scene *scene, double x, double y)
-{
-	t_vec3	*vec;
-	t_vec3	*camera_ray;
-	t_vec3	*u;
-	t_vec3	*v;
-
-	camera_ray = calc_camera_ray(scene);
-	u = calc_u(scene, camera_ray, x);
-	v = calc_v(scene, camera_ray, y);
-	vec = vec3_add4(scene->camera->point, camera_ray, u, v);
-	free(camera_ray);
-	free(u);
-	free(v);
-	return (vec);
-}
 
 t_window_info	*init_window_info(void)
 {
@@ -134,7 +27,8 @@ t_window_info	*init_window_info(void)
 	return (info);
 }
 
-t_object	*find_nearest_objects(t_ray *ray, t_list *objects, t_object *ignore_object)
+t_object	*find_nearest_objects(t_ray *ray, t_list *objects, \
+									t_object *ignore_object)
 {
 	t_object	*rtv;
 	double		t;
@@ -158,7 +52,8 @@ t_object	*find_nearest_objects(t_ray *ray, t_list *objects, t_object *ignore_obj
 	return (rtv);
 }
 
-bool	is_draw_shadow(t_ray *camera_ray, t_object *object, t_light *light, t_list *objects)
+bool	is_shadow(t_ray *camera_ray, t_object *object, \
+					t_light *light, t_list *objects)
 {
 	double		t;
 	t_vec3		*p_vec;
@@ -180,30 +75,42 @@ bool	is_draw_shadow(t_ray *camera_ray, t_object *object, t_light *light, t_list 
 	return (true);
 }
 
-void	draw(t_window_info *info, t_scene *scene)
+int	calc_pixel_color(int x, int y, t_scene *sc)
 {
-	int			i;
-	int			j;
-	t_ray		ray;
 	t_object	*object;
 	t_vec3		*vec_3d;
+	t_ray		ray;
+	int			color;
 
-	ray.start_vec = scene->camera->point;
+	color = COLOR_BLACK;
+	ray.start_vec = sc->camera->point;
+	vec_3d = to_screen_vec(sc, x, y);
+	ray.direction_vec = vec3_sub(vec_3d, ray.start_vec);
+	object = find_nearest_objects(&ray, sc->objects, NULL);
+	if (object != NULL && !is_shadow(&ray, object, sc->light, sc->objects))
+	{
+		color = add_color(color, calc_diffuse_light(&ray, object, sc->light));
+		color = add_color(color, calc_ambient_light(sc->ambient_light, object));
+	}
+	free(vec_3d);
+	free(ray.direction_vec);
+	return (color);
+}
+
+void	draw(t_window_info *info, t_scene *scene)
+{
+	int	i;
+	int	j;
+	int	color;
+
 	i = 0;
 	while (i < W)
 	{
 		j = 0;
 		while (j < H)
 		{
-			vec_3d = to_3d(scene, i, j);
-			ray.direction_vec = vec3_sub(vec_3d, ray.start_vec);
-			object = find_nearest_objects(&ray, scene->objects, NULL);
-			if (object != NULL && is_draw_shadow(&ray, object, scene->light, scene->objects) == false)
-			{
-				pixel_put_to_image(info->img, i, j, add_color(calc_diffuse_light(&ray, object, scene->light), calc_ambient_light(scene->ambient_lightning, object)));
-			}
-			free(vec_3d);
-			free(ray.direction_vec);
+			color = calc_pixel_color(i, j, scene);
+			pixel_put_to_image(info->img, i, j, color);
 			j++;
 		}
 		i++;
